rename solver card
